dashboard/requirements.txt
dash==3.3.0
pandas==2.3.3
mysql-connector-python==9.5.0
gunicorn==23.0.0
